Sat, Jul 20, 2024, 7:03 PM +03:30

+++ average.c
@@ -1,7 +1,7 @@
 #include <cs50.h>
 #include <stdio.h>
 
-int sum(int vary,float scores[]);
+int sum(int vary,int scores[]);
 int main(void)
 {
     int l = get_int("how meny number you have :");
@@ -13,7 +13,7 @@ int main(void)
     int g = sum(l , score);
     printf("your average is %f\n", g / (float) l );
 }
-int sum(int vary,float scores[])
+int sum(int vary,int scores[])
 {
     int x=0;
     for(int i = 0; i < vary; i++)
